clean code

+++ mindspore/core/ops/sequence_add_offset.cc
@@ -40,7 +40,7 @@ AbstractBasePtr SequenceAddOffsetInferInner(const PrimitivePtr &primitive,
                                             const std::vector<AbstractBasePtr> &input_args) {
   MS_EXCEPTION_IF_NULL(primitive);
   auto prim_name = primitive->name();
-  constexpr size_t input_len = 2;
+  constexpr int64_t input_len = 2;
   (void)CheckAndConvertUtils::CheckInteger("input number", SizeToLong(input_args.size()), kEqual, input_len, prim_name);
   abstract::AbstractBasePtrList abs{};
   abs.push_back(std::make_shared<abstract::AbstractScalar>(kValueAny, kInt64));
